--- main/templates/story.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>{{.Title}}</title>
-</head>
-<body>
-    <h1>{{.Title}}</h1>
-    <div class="story-content">
-        {{range .Story}}
-            <p>{{.}}</p>
-        {{end}}
-    </div>
-    <div class="variations-links">
-        {{range .Options}}
-            <div class="link-container">
-                <a href="{{.Arc}}">{{.Text}}</a>
-            </div>
-        {{end}}
-    </div>
-</body>
-</html>
