+++ students/cherednichenkoa/main.go
@@ -2,12 +2,12 @@ package main
 
 import (
 	"flag"
-	"gopherex/cyoa/main/route-handler"
-	"gopherex/cyoa/main/settings"
+	"gopherex/cyoa/students/cherednichenkoa/route-handler"
+	"gopherex/cyoa/students/cherednichenkoa/settings"
 )
 
 const (
-	storyTemplate = "main/templates/story.html"
+	storyTemplate = "cherednichenkoa/templates/story.html"
 )
 
 var (
@@ -17,6 +17,9 @@ var (
 
 func main()  {
 	flag.Parse()
+	if len(*listenPort) == 0 || len(*filePath) == 0 {
+		panic("Please specify application params (listenPort and filePath)")
+	}
 	config := settings.Settings{FilePath: *filePath, ListenPort: *listenPort, TemplatePath: storyTemplate}
 	handler := route_handler.RouteHandler{Settings: config}
 	handler.ServeRequests()
