+++ students/cherednichenkoa/templates/story.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>{{.Title}}</title>
+</head>
+<body>
+    <h1>{{.Title}}</h1>
+    <div class="story-content">
+        {{range .Story}}
+            <p>{{.}}</p>
+        {{end}}
+    </div>
+    <div class="variations-links">
+        <ul>
+            {{if .Options }}
+                {{range .Options}}
+                    <li class="link-container">
+                        <a href="{{.Arc}}">{{.Text}}</a>
+                    </li>
+                {{end}}
+            {{else}}
+                <p ><b>The End !</b></p>
+            {{end}}
+
+
+        </ul>
+    </div>
+</body>
+</html>
